chore: refactor props
