+++ components/run-info/map/map.js
@@ -1,4 +1,4 @@
-import { data, props } from "/options/props/map";
+import { data, props } from "/options/props/run-info/map";
 import observers from "./observers";
 import { lifeHanders, methods } from "./handlers";
