+++ components/run-info/overview/overview.js
@@ -1,6 +1,6 @@
 import { methods, lifeHanders } from "./handlers";
 import observers from "./observers";
-import { data, props } from "/options/props/overview";
+import { data, props } from "/options/props/run-info/overview";
 
 Component({
   options: {
