+++ components/run-info/tabs/tabs.js
@@ -1,5 +1,5 @@
 import methods from "./handlers";
-import { data, props } from "/options/props/tabs";
+import { data, props } from "/options/props/run-info/tabs";
 
 Component({
   data,
