+++ options/props/run-info/map.js
@@ -1,4 +1,4 @@
-import { DEFAULT_POSITION, DEFAULT_STATION, NOP } from "./defaults";
+import { DEFAULT_POSITION, DEFAULT_STATION, NOP } from "../defaults";
 
 export const data = {
   stationMarkers: [],
